+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changes For IRIS Service
 unreleased
 ==========
 
+ - EP /token/<token>/petitions: Return 'status' with 'response_token_usable' or
+   'response_token_used'
+
 2017/02/07 0.7.5
 ================